Fixes unparsed colours in profile viewer
